+++ src/server/ContactSystem.Core/Domain.Core/Queries/Interfaces/IPaginationQuery.cs
@@ -2,7 +2,7 @@ namespace ContactSystem.Core.Domain.Core.Queries.Interfaces;
 
 public interface IPaginationQuery
 {
-	long? Offset { get; }
+	int? Offset { get; }
 
-	long? Limit { get; }
+	int? Limit { get; }
 }
