+++ src/server/ContactSystem.Core/Domain/Pagination/Common/Extensions/EnumerableExtensions.cs
@@ -4,6 +4,6 @@ using Domain.Pagination;
 
 public static class EnumerableExtensions
 {
-	public static PagedList<T> ToPagedList<T> ( this IEnumerable<T> collection , long count , long offset , long limit )
+	public static PagedList<T> ToPagedList<T> ( this IEnumerable<T> collection , int count , int offset , int limit )
 		=> PagedList<T>.Create ( collection , count , offset , limit );
 }
